backend/requirements.txt
# FastAPI and server
fastapi[all]==0.109.0
uvicorn[standard]==0.27.0

# Database
supabase==2.3.4
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9

# AI
openai==1.10.0

# Configuration
pydantic==2.6.0
pydantic-settings==2.1.0
python-dotenv==1.0.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6

# Utilities
httpx==0.26.0
python-dateutil==2.8.2
tenacity==8.2.3

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
